[Finish] chapter10

+++ src/chapter10/Wand.java
@@ -3,9 +3,7 @@ package chapter10;
 public class Wand {
 	private String name;
 	private double power;
-	public String getName() {
-		return this.name;
-	}
+	public String getName() { return this.name; }
 	public void setName(String name) {
 		if(name == null || name.length() < 3) {
 			throw new IllegalArgumentException("長さが短すぎるか入力されていません。");	
@@ -13,11 +11,9 @@ public class Wand {
 			this.name = name;
 		}
 	}
-	public double getPower() {
-		return this.power;
-	}
-	public void setPower(int power) {
-		if(power < 0.5 || power > 100) {
+	public double getPower() { return this.power; }
+	public void setPower(double power) {
+		if(power < 0.5 || power > 100.0) {
 			throw new IllegalArgumentException("魔力の値が異常です。");	
 		} else {
 			this.power = power;
